Debug 消息使用英文

+++ src/main/java/org/mizoguchi/misaki/pojo/vo/front/AssistantFrontResponse.java
@@ -31,9 +31,9 @@ public class AssistantFrontResponse {
 
     private Long creatorId;
 
-    private Long likes;
+    private Integer likes;
 
-    private Long duplicateName;
+    private Integer duplicateName;
 
     private Boolean publicFlag;
